requirements.txt
langchain_openai
langserve
fastapi
uvicorn
langchain_community
sse_starlette
streamlit